backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# HTTP Client
httpx==0.28.1

# TextWorld
textworld==1.6.2

# Google Gemini AI
google-generativeai==0.8.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Utilities
python-multipart==0.0.20

